fix console logs and rename file
