+++ modules/app/components/SystemStatsSidebar.tsx
@@ -11,7 +11,7 @@ import { useDaiSavingsRate } from 'modules/web3/hooks/useDaiSavingsRate';
 import { useTokenBalance } from 'modules/web3/hooks/useTokenBalance';
 import { useMkrOnHat } from 'modules/web3/hooks/useMkrOnHat';
 import { formatValue } from 'lib/string';
-import { useContractAddress } from 'modules/web3/hooks/useChiefContract';
+import { useContractAddress } from 'modules/web3/hooks/useContractAddress';
 import { getEtherscanLink } from 'modules/web3/helpers/getEtherscanLink';
 
 type StatField =
