+++ sentry.client.config.js
@@ -9,7 +9,8 @@ const SENTRY_DSN = process.env.SENTRY_DSN || process.env.NEXT_PUBLIC_SENTRY_DSN;
 Sentry.init({
   dsn: SENTRY_DSN || 'https://a5f452e07e714442a31a16ecd9b42eac@o1007677.ingest.sentry.io/5970673',
   // Adjust this value in production, or use tracesSampler for greater control
-  tracesSampleRate: 1.0
+  tracesSampleRate: 1.0,
+  integrations: [new Sentry.Integrations.Breadcrumbs({ console: false })]
   // ...
   // Note: if you want to override the automatic release value, do not set a
   // `release` value here - use the environment variable `SENTRY_RELEASE`, so
